Updated to 1.19.4!~

+++ src/main/java/unlimited/litehacks/mods/movement/NoFall.java
@@ -27,7 +27,7 @@ public class NoFall extends Module {
             }
             if (mc.player.fallDistance > 2.5f && NoFallMode.isMode("packet")) {
                 if (mc.world.getBlockState(mc.player.getBlockPos().add(
-                        0, -1.5 + (mc.player.getVelocity().y * 0.1), 0)).getBlock() != Blocks.AIR) {
+                        0, (int) (-1.5 + (mc.player.getVelocity().y * 0.1)), 0)).getBlock() != Blocks.AIR) {
                     mc.player.networkHandler.sendPacket(new PlayerMoveC2SPacket.OnGroundOnly(false));
                     mc.player.networkHandler.sendPacket(new PlayerMoveC2SPacket.PositionAndOnGround(
                             mc.player.getX(), mc.player.getY() - 420.69, mc.player.getZ(), true));
